+++ R/app_label.R
@@ -7,8 +7,8 @@
 #'  \item Create database -- if no database exists to store the annotations,
 #'  use this button to create one
 #'  \item Choose database -- choose the database to store the annotations
-#'  \item Butterworth filter -- indicate low and high frequencies in kHz to 
-#'  filter the recordings
+#'  \item Butterworth filter -- check box to apply filter and 
+#'  indicate low and high frequencies in kHz to filter the recordings
 #'  \item Time expanded -- only used in recorders specifically intended for
 #'  bat recordings. Can take any numeric value. If the recording is not time
 #'  expanded the value must be set to 1. If it is time expanded the numeric
@@ -102,10 +102,8 @@ app_label <- function() {
           width = "100%"
         ),
         htmltools::hr(),
-        htmltools::h5(
-          "Butterworth filter (kHz)",
-          align = "center"
-        ),
+         shiny::checkboxInput("but_filt", "Butterworth filter (kHz)",
+                              value = FALSE, width = NULL),
         shiny::fluidRow(
           shiny::column(
             width = 6,
@@ -113,7 +111,7 @@ app_label <- function() {
             shiny::textInput(
               inputId = "low",
               label = "Low",
-              value = "10"
+              value = "0"
             )
           ),
           shiny::column(
@@ -133,7 +131,7 @@ app_label <- function() {
           shiny::numericInput(
             inputId = "tx",
             label = "Time expanded",
-            value = "10"
+            value = "1"
           )
           )
         ),
@@ -147,8 +145,8 @@ app_label <- function() {
         htmltools::hr(),
         shinyFiles::shinyDirButton(
           id = "noise_folder",
-          label = "Add non-relevant",
-          title = "Add non-relevant",
+          label = "Add irrelevant",
+          title = "Add irrelevant",
           style = "width:100%"
         )
       ),
@@ -228,9 +226,16 @@ app_label <- function() {
                   choices = c(
                     "1 ms" = "1",
                     "2 ms" = "2",
-                    "3 ms" = "3",
                     "4 ms" = "4",
-                    "5 ms" = "5"
+                    "10 ms" = "10",
+                    "20 ms" = "20",
+                    "40 ms" = "40",
+                    "100 ms" = "100",
+                    "200 ms" = "200",
+                    "400 ms" = "400",
+                    "1000 ms" = "1000",
+                    "2000 ms" = "2000",
+                    "4000 ms" = "4000"
                   ),
                   selected = "5"
                 )
@@ -243,12 +248,13 @@ app_label <- function() {
                   inputId = "timeStep",
                   label = "Overlap",
                   choices = c(
+                    "50%" = "0.5",
                     "60%" = "0.4",
                     "70%" = "0.3",
                     "80%" = "0.2",
                     "90%" = "0.1"
                   ),
-                  selected = "0.2"
+                  selected = "0.5"
                 )
               ),
               shiny::column(
@@ -261,7 +267,7 @@ app_label <- function() {
                     "medium resolution" = "4",
                     "high resolution" = "8"
                   ),
-                  selected = "4"
+                  selected = "1"
                 )
               )
             ),
@@ -413,7 +419,7 @@ Spectrogram visualization:
 
       sound <- import_audio(
         path = input$files,
-        butt = TRUE,
+        butt = input$but_filt,
         low = as.numeric(input$low),
         high = as.numeric(input$high),
         tx = as.numeric(input$tx)
@@ -572,6 +578,7 @@ Spectrogram visualization:
           sound <-
             import_audio(
               path = paste0(folder_path, "/", noise_files[i]),
+              butt = input$but_filt,
               low = as.numeric(input$low),
               high = as.numeric(input$high)
             )